changed in controller structure for consdering technology , based on technology  it will rediirect to respective controller and methods
