--- app/Http/Controllers/BusinessModels/ContentWritingController.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\BusinessModels;
-
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-
-class ContentWritingController extends Controller
-{
-    //
-}
